[INLONG-7908][Sort] Update cdc-base for Incremental Snapshot Framework
