+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/source/meta/split/FinishedSnapshotSplitInfo.java
@@ -18,6 +18,7 @@
 package org.apache.inlong.sort.cdc.base.source.meta.split;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
+import static org.apache.inlong.sort.cdc.base.util.RecordUtils.shouldUseCatalogBeforeSchema;
 
 import io.debezium.relational.TableId;
 import java.io.IOException;
@@ -28,7 +29,7 @@ import org.apache.flink.util.FlinkRuntimeException;
 import org.apache.inlong.sort.cdc.base.source.meta.offset.Offset;
 import org.apache.inlong.sort.cdc.base.source.meta.offset.OffsetDeserializerSerializer;
 import org.apache.inlong.sort.cdc.base.source.meta.offset.OffsetFactory;
-import org.apache.inlong.sort.cdc.base.util.SerializerUtils;
+import com.ververica.cdc.connectors.base.utils.SerializerUtils;
 
 /** The information used to describe a finished snapshot split.
  * Copy from com.ververica:flink-cdc-base:2.3.0.
@@ -139,7 +140,7 @@ public class FinishedSnapshotSplitInfo implements OffsetDeserializerSerializer {
 
     public byte[] serialize(final DataOutputSerializer out) throws IOException {
         boolean useCatalogBeforeSchema =
-                SerializerUtils.shouldUseCatalogBeforeSchema(this.getTableId());
+                shouldUseCatalogBeforeSchema(this.getTableId());
         out.writeBoolean(useCatalogBeforeSchema);
         out.writeUTF(this.getTableId().toString());
         out.writeUTF(this.getSplitId());
