+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/source/meta/split/SourceSplitSerializer.java
@@ -38,7 +38,9 @@ import org.apache.inlong.sort.cdc.base.source.meta.offset.Offset;
 import org.apache.inlong.sort.cdc.base.source.meta.offset.OffsetDeserializerSerializer;
 import org.apache.inlong.sort.cdc.base.source.meta.offset.OffsetFactory;
 import org.apache.inlong.sort.cdc.base.source.meta.split.MetricSplit.TableMetric;
-import org.apache.inlong.sort.cdc.base.util.SerializerUtils;
+import com.ververica.cdc.connectors.base.utils.SerializerUtils;
+
+import static org.apache.inlong.sort.cdc.base.util.RecordUtils.shouldUseCatalogBeforeSchema;
 
 /** A serializer for the {@link SourceSplitBase}.
  * Copy from com.ververica:flink-cdc-base:2.3.0.
@@ -73,7 +75,7 @@ public abstract class SourceSplitSerializer
             final DataOutputSerializer out = SERIALIZER_CACHE.get();
             out.writeInt(SNAPSHOT_SPLIT_FLAG);
             boolean useCatalogBeforeSchema =
-                    SerializerUtils.shouldUseCatalogBeforeSchema(snapshotSplit.getTableId());
+                    shouldUseCatalogBeforeSchema(snapshotSplit.getTableId());
             out.writeBoolean(useCatalogBeforeSchema);
             out.writeUTF(snapshotSplit.getTableId().toString());
             out.writeUTF(snapshotSplit.splitId());
@@ -204,8 +206,7 @@ public abstract class SourceSplitSerializer
         final int size = tableSchemas.size();
         out.writeInt(size);
         for (Map.Entry<TableId, TableChange> entry : tableSchemas.entrySet()) {
-            boolean useCatalogBeforeSchema =
-                    SerializerUtils.shouldUseCatalogBeforeSchema(entry.getKey());
+            boolean useCatalogBeforeSchema = shouldUseCatalogBeforeSchema(entry.getKey());
             out.writeBoolean(useCatalogBeforeSchema);
             out.writeUTF(entry.getKey().toString());
             final String tableChangeStr =
