+++ inlong-sort/sort-connectors/cdc-base/src/main/java/org/apache/inlong/sort/cdc/base/util/RecordUtils.java
@@ -60,6 +60,9 @@ public class RecordUtils {
     public static final String CONNECTOR = "connector";
     public static final String MYSQL_CONNECTOR = "mysql";
 
+    public static final String SCHEMA_HEARTBEAT_EVENT_KEY_NAME =
+            "io.debezium.connector.common.Heartbeat";
+
     private RecordUtils() {
 
     }
@@ -160,6 +163,18 @@ public class RecordUtils {
 
     public static boolean isDdlRecord(Struct value) {
         return value.schema().field(HISTORY_RECORD_FIELD) != null;
+    }
+
+    public static boolean isHeartbeatEvent(SourceRecord record) {
+        Schema valueSchema = record.valueSchema();
+        return valueSchema != null
+                && SCHEMA_HEARTBEAT_EVENT_KEY_NAME.equalsIgnoreCase(valueSchema.name());
+    }
+
+    public static boolean shouldUseCatalogBeforeSchema(TableId tableId) {
+        // if catalog is not defined but the schema is defined return this flag as false
+        // otherwise return true
+        return !(tableId.catalog() == null && tableId.schema() != null);
     }
 
 }
